Fixes for initial testing
