+++ Discord Bot/.gitignore
@@ -1,2 +1,2 @@
 test.py
-token.py
+config.py
